Fixed file names and moved require simplecov to spec_helper only;
