+++ .gitignore
@@ -1,3 +1,2 @@
 coverage/
-DS_Store
 .DS_Store
