+++ lib/market.rb
