+++ lib/sale.rb
