+++ lib/vendor.rb
