+++ specs/market_spec.rb
@@ -1,6 +1,3 @@
-require 'simplecov'
-SimpleCov.start
-
 require_relative 'spec_helper'
 
 describe FarMar::Market do
