+++ specs/product_spec.rb
@@ -1,0 +1,8 @@
+require_relative 'spec_helper'
+
+describe FarMar::Product do
+  it "exists" do
+    FarMar::Product.wont_be_nil
+  end
+
+end
