+++ specs/sale_spec.rb
@@ -1,0 +1,8 @@
+require_relative 'spec_helper'
+
+describe FarMar::Sale do
+  it "exists" do
+    FarMar::Sale.wont_be_nil
+  end
+
+end
