+++ specs/spec_helper.rb
@@ -1,3 +1,6 @@
+require 'simplecov'
+SimpleCov.start
+
 require 'minitest'
 require 'minitest/spec'
 require 'minitest/autorun'
