+++ specs/vendor_spec.rb
@@ -1,0 +1,8 @@
+require_relative 'spec_helper'
+
+describe FarMar::Vendor do
+  it "exists" do
+    FarMar::Vendor.wont_be_nil
+  end
+
+end
