--- specs/vendors_spec.rb
@@ -1,4 +1,0 @@
-require 'simplecov'
-SimpleCov.start
-
-require_relative 'spec_helper'
